+++ js/util.js
@@ -1,0 +1,13 @@
+function getRandomInt(min, max) {
+  min = Math.ceil(min);
+  max = Math.floor(max);
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+   }
+   getRandomInt(1, 25);
+  function checkStringLength(stringToCheck, maxLength) {
+    if (stringToCheck.length <= maxLength) {
+      return true;
+    }
+    return false;
+  }
+  checkStringLength('book', 5);
